Run "prettier" on all js/ts code and enforce on Travis.
We use a width of 100. This allows people to easily fit two versions of
the code (or different editors) side by side on one screen. More than
this is tricky to fit (e.g. in github diff views) or in terminals. Less
than this (e.g. 80) introduces a fair lot more newlines that don't help
readability a lot.